Delete command won't fail if a team without a user already exists

+++ kos/management/commands/deleteusers.py
@@ -11,13 +11,15 @@ class Command(BaseCommand):
         parser.add_argument('game_id', type=int)
 
     def handle(self, *args, **options):
-        game_id:int = options['game_id']
+        game_id: int = options['game_id']
         teams = Team.objects.filter(game=game_id).all()
         for team in teams:
-            team.email = team.user.email
             user = team.user
+            if user is None:
+                print(f'Team {team} already has no user')
+                continue
+            team.email = user.email
             print(f'Deleting user {user}')
             team.user = None
             user.delete()
             team.save()
-
